+++ code/chap01/scala/src/main/scala/org/data/algorithms/spark/ch01/AggregateByKey.scala
@@ -1,4 +1,4 @@
-package com.spark.algos.data.ch1
+package org.data.algorithms.spark.ch01
 
 import org.apache.spark.sql.SparkSession
